requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
alembic==1.13.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
dataclasses-json==0.6.7
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0.post1
environs==9.5.0
fastapi==0.109.1
filelock==3.18.0
frozenlist==1.6.2
fsspec==2025.5.1
google-ai-generativelanguage==0.4.0
google-api-core==2.25.0
google-auth==2.40.3
google-generativeai==0.4.1
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwipiepy==0.20.3
kiwipiepy_model==0.20.0
langchain==0.1.12
langchain-community==0.0.38
langchain-core==0.1.53
langchain-google-genai==0.0.11
langchain-text-splitters==0.0.2
langdetect==1.0.9
langsmith==0.1.147
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
minio==7.2.15
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
networkx==3.5
numpy==1.26.4
orjson==3.10.18
packaging==23.2
pandas==2.3.0
passlib==1.7.4
pillow==11.2.1
portalocker==2.10.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.8
psycopg2-binary==2.9.9
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pymilvus==2.3.6
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
pytz==2025.2
pywin32==310
PyYAML==6.0.2
qdrant-client==1.14.2
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==2.5.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.27
starlette==0.35.1
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.27.0
yarl==1.20.0
